my-money-backend/pyproject.toml
[tool.poetry]
name = "my-money-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
mixer = "^7.2.0"
pytest-django = "^4.5.2"
model-bakery = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
